Change singular board semantics
Previously, we would call boardgroup.singular() in the Robot
constructor, meaning that one could not instantiate a Robot
object unless exactly one of every board is connected.

After this change, a Robot can now be instantiated in this
situation. It is only when you try to access the relevant
properties (e.g. motor_board) of the Robot object that an
exception will be thrown.