Improve timeout handling (#326)
* Use Timer rather than SIGALARM

This means the implementation is cross-platform. Because it's a thread timer, it shouldn't have the same issues as `time.sleep`

* Kill the process manually

`interrupt_main` sends a `SIGINT`, which can trivially be caught by bad code. `SIGTERM` can't.

Tested working on Windows

* Add an E2E test for timeout

Because of the nature, a monkeypatched unittest isn't quite enough.

* Return successfully in the event of a timeout

* Don't catch signals on non-linux

It's not possible to distinguish non-timeout terminations on Windows, so we don't. We don't care about it on these platforms anyway.

* Remove ternary for platform check

Mypy doesn't like it

* Use SIGALRM on Unix, and a timer everywhere else

This keeps the production implementation known-working and simple, and approximates it on Windows.

This also has the benefit of ensuring a timeout yields a successful return code on the robot.

* Add specific test for Windows